src/main/java/me/motyim/hackerrank/algorithm/warmup/RepeatedString.java
package me.motyim.hackerrank.algorithm.warmup;

import java.io.BufferedWriter;
import java.io.FileWriter;
import java.io.IOException;
import java.util.Scanner;

public class RepeatedString {
    // Complete the repeatedString function below.
    static long repeatedString(String s, long n) {
        long count = getCountA(s);
        long addtionCount = getCountA(s.substring(0, (int) (n % s.length())));
        return count * (n-n % s.length()) / s.length() + addtionCount;
    }

    private static long getCountA(String s) {
        return s.chars().filter(c -> c == 'a').count();
    }


    private static final Scanner scanner = new Scanner(System.in);

    public static void main(String[] args) throws IOException {
        BufferedWriter bufferedWriter = new BufferedWriter(new FileWriter(System.getenv("OUTPUT_PATH")));

        String s = scanner.nextLine();

        long n = scanner.nextLong();
        scanner.skip("(\r\n|[\n\r\u2028\u2029\u0085])?");

        long result = repeatedString(s, n);

        bufferedWriter.write(String.valueOf(result));
        bufferedWriter.newLine();

        bufferedWriter.close();

        scanner.close();
    }
}
